Add files via upload
code for generating patches from fundus image

patches are collected from DRIVE dataset - 1000 patch from each image - total 20000 train and 20000 test patches. this are stored in DRIVEtrainData and DRIVEtestData - the file is in npy format which stores numpy array

DataLocation and DataLabels contain the location (row, column of the middle pixel of the patch) and label (whether the middle pixel is vessel or not). 

the dataset is uploaded in google drive